refactor: MockStorage, HandlerResponse and everything related to it has been moved to mocks/utils.go.
